A3SAT.Animation.Primitives

Description

Documentation

showSvg :: [Text] -> LExpression l -> LExpression (l, SVG)

Render the boolean expression with LaTeX, then associate nodes with glyphs.

showSvg' :: [Text] -> LExpression l -> LExpression SVG

Render the boolean expression with LaTeX, drop the original labels, then associate nodes with glyphs.

mapVar :: ASetter' l b -> [a] -> (Word -> Bool) -> (a -> b -> b) -> LExpression l -> LExpression l

Modify the variables.

mapVar idx attrs filt upd does the following:

Locate all the variables using vars, filtered by filt

Focus on the targeted parts with the given Lens' idx

Modify that part with upd, provided the corresponding attribute in attrs, indexed by the Word index of that variable

colourVar :: Lens' l SVG -> [Texture] -> LExpression l -> LExpression l

Colour the variables.

colourLabel :: ASetter' (LExpression l) SVG -> Texture -> LExpression l -> LExpression l

Colour labels by category.

exprSvg :: LExpression SVG -> SVG

Collect all SVGs in an expression.

morphAnim :: LExpression (SVG, SVG) -> Animation

Perform linear morphology on already-paired-up images.

morphAnimZip :: LExpression SVG -> LExpression SVG -> Animation

Perform linear morphology on a pair of expressions.

evalExpr :: [Bool] -> LExpression l -> Bool

Evaluate a boolean expression.

evalExprIdx :: Applicative f => Lens' l (f Bool) -> LExpression l -> LExpression l

Evaluate a boolean expression, with its label including boolean values.

evalExprF :: Applicative f => [f Bool] -> LExpression l -> f Bool

Evaluate a boolean expression in an Applicative environment. Provided a list of boolean values.

evalExprWith :: Applicative f => (l -> Word -> f Bool) -> LExpression l -> f Bool

Evaluate a boolean expression in an Applicative environment. Provided a function from labels and indices to boolean values.

evalExprAlg :: Applicative f => (l -> Word -> f Bool) -> ExpressionF (f Bool) l -> f Bool

Algebra ExpressionF (f Bool) l -> f Bool for evaluating a boolean expression.

exprBoundingBox :: LExpression SVG -> (Double, Double, Double, Double)

Bounding box for expressions.

mkBoundingRect :: LExpression SVG -> SVG

Make a bounding rectangle for the expression.