A3SAT.Expression

Description

Expression Functors

Expression Functors with Labels

data ExpressionF a l

The endofunctor for boolean expressions.

Constructors

VarF l Word

NotF l a

AndF a l a

OrF a l a

ParenF l a l

glyphCount :: [Int] -> ExpressionF a l -> Int

Given number of glyphs for variables, get the number of glyphs for this element.

Expression Constructors with Labels

pattern LVar :: l -> Word -> LExpression l

Labeled Variable.

pattern LNot :: l -> LExpression l -> LExpression l

Boolean negation with label.

pattern LAnd :: LExpression l -> l -> LExpression l -> LExpression l

Boolean and with label.

pattern LOr :: LExpression l -> l -> LExpression l -> LExpression l

Boolean or with label.

pattern LParen :: l -> LExpression l -> l -> LExpression l

Parenthesis with left and right labels.

Expression Constructors without Labels

pattern Var :: Word -> Expression

Variable with a dummy label.

pattern Not :: Expression -> Expression

Boolean negation with a dummy label.

pattern And :: Expression -> Expression -> Expression infixr 3

Boolean and with a dummy label.

pattern Or :: Expression -> Expression -> Expression infixr 2

Boolean or with a dummy label.

pattern Paren :: Expression -> Expression

Parenthesis with dummy labels.

Boolean Expressions

type LExpression = Fix ExpressionF

Boolean expressions with labels.

type Expression = LExpression ()

Boolean expressions.

Utility Functions

addParenWith :: (l, l) -> LExpression l -> LExpression l

Add necessary parentheses for the boolean expression. Parentheses labeled with the provided label.

addParen :: Expression -> Expression

Add necessary parentheses for the boolean expression.

mapExprM

Arguments

:: Applicative f

=> (forall a b. ExpressionF a b -> c -> f d)
Expression visitor: expression type -> old label -> new label.

-> LExpression c

-> f (LExpression d)

Transform the expression with the given visitor.

zipExprFWith :: (a -> a' -> a'') -> (l -> l' -> l'') -> (ExpressionF a l -> ExpressionF a' l' -> ExpressionF a'' l'') -> ExpressionF a l -> ExpressionF a' l' -> ExpressionF a'' l''

Zip 2 expression functors with of same shape with given functions.

zipExprF :: ExpressionF a l -> ExpressionF a' l' -> ExpressionF (a, a') (l, l')

Zip 2 expression functors of the same shape.

zipExprWith :: (l -> l' -> l'') -> (LExpression l -> LExpression l' -> LExpression l'') -> LExpression l -> LExpression l' -> LExpression l''

Zip 2 expressions of the same shape with given function.

zipExpr :: LExpression l -> LExpression l' -> LExpression (l, l')

Zip 2 expressions of the same shape with given function.