A3SAT.Expression.Lens

Description

Optics for Expression Functor

_ParenF :: forall a l. Prism' (ExpressionF a l) (l, a, l)

_OrF :: forall a l. Prism' (ExpressionF a l) (a, l, a)

_AndF :: forall a l. Prism' (ExpressionF a l) (a, l, a)

_NotF :: forall a l. Prism' (ExpressionF a l) (l, a)

_VarF :: forall a l. Prism' (ExpressionF a l) (l, Word)

_BinOpF :: Traversal' (ExpressionF a l) (a, l, a)

Prisms for AndF and OrF.

varF :: (Word -> Bool) -> Traversal' (ExpressionF a l) (l, Word)

Label for variables matching the predicate.

allVarF :: Traversal' (ExpressionF a l) l

Label for all variables.

notF :: Traversal' (ExpressionF a l) l

Label for NotF.

andF :: Traversal' (ExpressionF a l) l

Label for AndF.

orF :: Traversal' (ExpressionF a l) l

Label for OrF.

binOpF :: Traversal' (ExpressionF a l) l

Label for AndF and OrF.

leftParenF :: Traversal' (ExpressionF a l) l

Left parentheses.

rightParenF :: Traversal' (ExpressionF a l) l

Right parentheses.

parenF :: Traversal' (ExpressionF a l) l

Both parentheses.

Optics for Expressions

vars :: (Word -> Bool) -> Traversal' (LExpression l) (l, Word)

Label for variables matching the predicate.

allVars :: Traversal' (LExpression l) l

Label for all variables.

nots :: Traversal' (LExpression l) l

Label for NotF.

ands :: Traversal' (LExpression l) l

Label for AndF.

ors :: Traversal' (LExpression l) l

Label for OrF.

binOps :: Traversal' (LExpression l) l

Label for AndF and OrF.

leftParens :: Traversal' (LExpression l) l

Left parentheses.

rightParens :: Traversal' (LExpression l) l

Right parentheses.

parens :: Traversal' (LExpression l) l

Both parentheses.

Global Optics on Expressions

transParen :: Traversal' (LExpression l) (LExpression l)

Travel through a parenthesis.

asProduct :: Traversal' (LExpression l) (LExpression l)

An LExpression as a product of a list of LExpressions.

asSum :: Traversal' (LExpression l) (LExpression l)

An LExpression as a sum of a list of LExpressions.

primaryLabel :: Lens' (LExpression l) l

The "primary" label for an LExpression.

Utility Functions

disjointUnion :: Prism' s a -> Prism' s a -> Traversal' s a

Make a disjoint union of 2 Prism's.

fixExpr :: Applicative f => (forall a. LensLike' f (ExpressionF a l) b) -> LensLike' f (LExpression l) b

Fixpoint for lenses on the expression functor.

WARNING: recursive traversal are done after non-recursive ones. Thus traversal does not follow the field order in data declaration.