A3SAT.LaTeX

Description

Documentation

prettyShow :: [Text] -> LExpression a -> Text

Get human-readable text for a boolean expression.

prettyPrint :: [Text] -> LExpression a -> IO ()

Print human-readable text for a boolean expression to stdout.

showLaTeX :: [Text] -> LExpression a -> Text

Get LaTeX sequence for a boolean expression.

locateGlyphsM :: MonadState Int m => [Int] -> LExpression a -> m (LExpression (a, Int))

Given number of glyphs for variables, label the position of each element. This function runs in a State monad, keeping track of the total count of glyphs.

locateGlyphs :: [Int] -> LExpression a -> LExpression (a, Int)

Given number of glyphs for variables, label the position of each element.

associateGlyphsM :: MonadState [SVG] m => [Int] -> LExpression a -> m (LExpression (a, SVG))

Given number of glyphs for variables, associate each element with its glyphs. This function runs in a State monad, keeping track of the remaining glyphs.

associateGlyphs :: [SVG] -> [Int] -> LExpression a -> LExpression (a, SVG)

Given number of glyphs for variables, associate each element with its glyphs.